Tidy up wallet
